Improve coverage multifit
